requirements.txt
librosa~=0.10.0.post2
Pillow~=9.5.0
synthviz~=0.0.2
httplib2~=0.22.0
google-api-python-client~=2.86.0
oauth2client~=4.1.3
piano-transcription-inference~=0.0.5